Move download panel to src folder
