+++ src/panels/builder.py
@@ -1,0 +1,1 @@
+"""Panel builder."""
