--- src/ui/panel.py
@@ -1,1 +1,0 @@
-"""Hana3D Blender Panel class."""
